+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # HEAD
 
+-   Make JSON the default format when importing/exporting the database [PR#28](https://github.com/ziotom78/instrumentdb/pull/28)
+
 -   Format JSON records and highlight their syntax using [Highlight.js](https://highlightjs.org/) ([PR#27](https://github.com/ziotom78/instrumentdb/pull/27))
 
 -   Increase the size of the `metadata` field for data files from 8 kB to 32 kB